Code quality improvements: fix all warnings and clippy issues
Fixes:
- Fixed 13 compiler warnings (unused imports, dead code)
- Fixed 5 clippy warnings (box_default, large_enum_variant)
- Added #[cfg(feature = rectifier)] to compute_histogram
- Added #[allow(dead_code)] to SmartPipeline fields
- Boxed Controller enum variants for size optimization

All 691 tests pass, 0 clippy warnings

+++ cam-rust/cam-isp/src/profile.rs
@@ -552,7 +552,7 @@ impl PipelineProfile {
         if self.use_vignetting {
             info!("  vignetting: VignettingBlock");
             blocks.push(Box::new(crate::blocks::VignettingBlock::new_default(
-                target_width as u32,
+                target_width,
                 (target_width as f64 * 9.0 / 16.0) as u32,
             )));
         } else {
